Refact to project
What do you think about a refactored project? And the most important thing: It's working!
