--- src/URLHandler.java
@@ -1,9 +1,0 @@
-
-public class URLHandler {
-	
-	private static String path = "/org/fireworksimulator/";
-	
-	public static String getResource(String filename) {
-		return path + "res/" + filename;
-	}
-}
