+++ src/org/fireworksimulator/model/BurningParticle.java
@@ -1,3 +1,5 @@
+package org.fireworksimulator.model;
+
