+++ src/org/fireworksimulator/model/Firework.java
@@ -1,3 +1,4 @@
+package org.fireworksimulator.model;
