+++ src/org/fireworksimulator/model/Particle.java
@@ -1,3 +1,4 @@
+package org.fireworksimulator.model;
