+++ src/org/fireworksimulator/model/Streak.java
@@ -1,3 +1,4 @@
+package org.fireworksimulator.model;
