+++ src/org/fireworksimulator/res/Ass4FXML.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 
-<SplitPane dividerPositions="0.8010033444816054" orientation="VERTICAL" prefHeight="600.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="MyController">
+<SplitPane dividerPositions="0.8010033444816054" orientation="VERTICAL" prefHeight="600.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.fireworksimulator.controller.MyController">
    <items>
       <StackPane prefHeight="150.0" prefWidth="200.0">
          <children>
