+++ src/org/fireworksimulator/res/background.png
