+++ src/org/fireworksimulator/util/EmitterException.java
@@ -1,3 +1,4 @@
+package org.fireworksimulator.util;
