+++ src/org/fireworksimulator/util/EnvironmentException.java
@@ -1,3 +1,4 @@
+package org.fireworksimulator.util;
